[rename/move] render.js 파일명을 pages.js 로 변경하고 render 폴더 내부로 이동 - main.js: pages.js와 관련된 모듈의 경로 변경
